Tuned knockback, added head-hitting.
Changed explosions to be a ball collider.

Refactored knockback calculations to be within a single file.

Walk Pathfinders can now hit their head (this applies for one way platforms as well).

Fixed the bug where the fly pathfinder wouldn't take knockback while it was patrolling.

Fixed the bug where the autotargeting system wouldn't recognize 360 degrees as Right and would believe it is closest to DownRight.