Remade enemy spawning to use an event system instead
Now all that's needed to spawn a new enemy is to send an event telling the system where it is, what stats to give it, and how many coins it has.

An all-new way to select difficulty was also added: EnemyDifficulty. This can be added within LDtk, and maps to different stats for different enemies.

Randomness is still there in the form of a range that's an extra parameter in the event struct.