Finished bossfight
All the animations are now in place. There was a bug with the hover that was fixed (still have no clue why it was caused, but it's fine now, soo...)

Also put some band-aids on the unsolved bugs. For the enemies flying OOB, they now should teleport back in. For the enemy colliders getting displaced, I center them every frame. For the player's collider working, I turn it off every frame they do not have the dash.

These are all temporary solutions but I don't want to waste the time finding the bug sources so I'll just do this for now.